backend/ingest/requirements.txt
pandas==2.2.2
requests
kafka-python==2.0.2
feedparser
python-telegram-bot==20.7
tweepy==4.14.0
telethon
